requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
bleach==6.2.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.3.1
colorama==0.4.6
cryptography==46.0.3
distro==1.9.0
fastapi==0.115.5
greenlet==3.3.0
groq==1.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.13.0
kaggle==1.6.17
numpy==2.2.2
openai==2.26.0
pandas==2.2.3
pycparser==2.23
pydantic==2.9.2
pydantic-settings==2.6.1
pydantic_core==2.23.4
pyotp==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-slugify==8.0.4
pytz==2025.1
requests==2.32.3
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.36
starlette==0.41.3
text-unidecode==1.3
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
webencodings==0.5.1